Split up server files
